+++ Solution/ContosoProject/Data/DummyData/DummyDAOForCategory.cs
@@ -1,0 +1,10 @@
+﻿using Domain.DAO;
+using Domain.Entities.Products;
+
+namespace Data.DummyData
+{
+    class DummyDAOForCategory : DummyDAO<Category>, ICategoryRepository
+    {
+        
+    }
+}
