+++ Solution/ContosoProject/Data/DummyData/DummyDAOForComment.cs
@@ -1,0 +1,9 @@
+﻿using Domain.DAO;
+using Domain.Entities.Comments;
+
+namespace Data.DummyData
+{
+    class DummyDAOForComment : DummyDAO<Comment>, ICommentRepository
+    {
+    }
+}
