--- Solution/ContosoProject/Data/EFData/EFDAOForCategory.cs
@@ -1,10 +1,0 @@
-﻿using Domain.DAO;
-using Domain.Entities.Products;
-
-namespace Data.EFData
-{
-    class EFDAOForCategory : EFDAO<Category>, ICategoryRepository
-    {
-        
-    }
-}
