--- Solution/ContosoProject/Data/EFData/EFDAOForComment.cs
@@ -1,9 +1,0 @@
-﻿using Domain.DAO;
-using Domain.Entities.Comments;
-
-namespace Data.EFData
-{
-    class EFDAOForComment : EFDAO<Comment>, ICommentRepository
-    {
-    }
-}
